Fix regex and exit on fswatch crashing.
On versions of fswatch < 1.12, the POSIX regex parser is used, which
doesn't crash on unnecessarily escaping a / character with a \. However,
in versions >= 1.12, the C++ regex parser is used, which is stricter
about this and crashes.

This meant that it worked fine on my machine and quietly failed for
other people.

Unison should complain if the fsmonitor process exits, so exiting the
main fsmonitor process on fswatch returning should be enough to make
this sort of error less quiet.